+++ src/models/match.h
@@ -17,7 +17,7 @@ class Match : public Model {
   // decreases.
   // description: a short identifier for this model.
   Match(ShortTermMemory& short_term_memory, LongTermMemory& long_term_memory,
-        unsigned long long table_size, const unsigned int& byte_context,
+        unsigned int table_size, const unsigned int& byte_context,
         int limit, std::string description, bool enable_analysis);
   void Predict(ShortTermMemory& short_term_memory,
                const LongTermMemory& long_term_memory);
